+++ lib/cocoapods_acknowledgements/addons/modifiers/metadata_plist_modifier.rb
@@ -6,6 +6,8 @@ module CocoaPodsAcknowledgements
   module AddOns
     class MetadataPlistModifier
 
+      # A modifier to update Pods/Pods-#{app_name}-metadata.plist.
+      #
       # @param target [Pod::Installer::PostInstallHooksContext::UmbrellaTargetDescription] the xcodeproj target.
       # @param sandbox [Pod::Sandbox] the CocoaPods sandbox
       #
@@ -34,19 +36,30 @@ module CocoaPodsAcknowledgements
         plist = CFPropertyList::List.new(file: @plist_path)
         entries = plist.value.value["specs"].value
         existing_titles = entries.map { |spec| spec.value["name"].value }
-        excluded_names += existing_titles
+        excluded_names.uniq!
 
-        additions = plist_metadata.map { |metadata|
-          next if metadata.nil? or excluded_names.include? metadata[:name]
+        additions = plist_metadata.map do |metadata|
+          next if metadata.nil? or existing_titles.include? metadata[:name]
           Pod::UI.info "Adding #{metadata[:name]} to #{@plist_path.basename}"
           CFPropertyList.guess(metadata)
-        }.reject(&:nil?)
+        end.reject(&:nil?)
 
         acknowledgements = entries + additions
-        acknowledgements.sort! { |a, b| a.value["name"].value <=> b.value["name"].value }
+        acknowledgements
+          .sort! { |a, b| a.value["name"].value <=> b.value["name"].value }
+          .reject! do |entry|
+            matches = excluded_names.any? do |excluded_name|
+              pattern = %r(^#{Regexp.escape(excluded_name).gsub("\*", ".*?")})
+              entry.value["name"].value =~ pattern
+            end
+            Pod::UI.info %(Removing #{entry.value["name"].value} from #{@plist_path.basename}) if matches
+            matches
+          end
 
         plist.value.value["specs"].value = acknowledgements
         plist.save(@plist_path, CFPropertyList::List::FORMAT_XML)
+
+        Pod::UI.puts "Saving #{@plist_path}".green
       end
 
     end